[Modified] redis 파일 따로 분리하였음
