+++ server/socket/index.ts
@@ -1,17 +1,5 @@
 import openSocket from 'socket.io';
-import redis from 'redis';
-import bluebird from 'bluebird';
-
-bluebird.promisifyAll(redis.RedisClient.prototype);
-bluebird.promisifyAll(redis.Multi.prototype);
-const port = process.env.REDIS_PORT as any;
-
-const redisConfig:redis.ClientOpts = {
-  host: process.env.REDIS_SERVER_ADDR as string,
-  port: port as number,
-  password: process.env.REDIS_PASSWORD as string,
-};
-const redisClient = redis.createClient(redisConfig);
+import { setUserSocketId } from '../redis';
 
 const connect = (server: any) => {
   const io = openSocket(server);
@@ -19,7 +7,7 @@ const connect = (server: any) => {
     socket.on('userInfo', (userInfo) => {
       const userId = userInfo.id;
       const socketId = socket.id;
-      redisClient.set(userId, socketId);
+      setUserSocketId(userId, socketId);
     });
   });
 };
